Feature/detailed quest label

+++ management/views.py
@@ -1,9 +1,20 @@
-from django.db.models import ProtectedError
-from rest_framework.exceptions import ValidationError
+from typing import cast
+
+from django.db.models import ProtectedError, Q, QuerySet
+from django.utils import timezone
+from rest_framework.exceptions import PermissionDenied, ValidationError
 from rest_framework.viewsets import ModelViewSet
 
+from users.models import Employee
+
 from .models import Activity, Quest
-from .serializers import ActivitySerializer, QuestCreatingSerializer, QuestSerializer
+from .serializers import (
+    ActivitySerializer,
+    QuestCreatingSerializer,
+    QuestSerializer,
+    QuestSimplifiedSerializer,
+    QuestUpdateReportSerializer,
+)
 
 
 class ActivityViewSet(ModelViewSet):
@@ -26,12 +37,36 @@ class QuestViewSet(ModelViewSet):
 
     queryset = Quest.objects.all()
 
+    def get_queryset(self) -> QuerySet:
+        """Ограничение набора отображаемых задач"""
+
+        user = cast(Employee, self.request.user)
+        if user.is_superuser:
+            return Quest.objects.all()
+        return Quest.objects.filter(Q(operator=user) | Q(creator=user))
+
     def get_serializer_class(self) -> type:
         """Определяет класс сериализатора в зависимости от
         совершаемого пользователем действия и его роли по отношению к задаче"""
 
         if self.action == "create":
             self.serializer_class = QuestCreatingSerializer
+        elif self.action == "list":
+            self.serializer_class = QuestSimplifiedSerializer
+        elif self.action in ["update", "partial_update"]:
+            quest = self.get_object()
+            if quest.dead_line < timezone.now():
+                raise PermissionDenied(
+                    "Задача не может быть изменена, после истечения времени, отведенного на ее выполнение"
+                )
+            user = self.request.user
+            if quest.creator == user:
+                self.serializer_class = QuestSerializer
+            else:
+                if quest.operator == user:
+                    self.serializer_class = QuestUpdateReportSerializer
+                else:
+                    raise PermissionDenied("Для изменения объекта задачи нужно быть ее создателем или исполнителем")
         else:
             self.serializer_class = QuestSerializer
         return self.serializer_class
@@ -39,6 +74,10 @@ class QuestViewSet(ModelViewSet):
     def perform_destroy(self, instance: Quest) -> None:
         """Исключает возможность удаления задачи, имеющей подзадачи"""
 
+        quest = self.get_object()
+        user = cast(Employee, self.request.user)
+        if not user.is_superuser and quest.creator != user:
+            raise PermissionDenied("Удалять задачи может только суперпользователь или их создатель")
         try:
             instance.delete()
         except ProtectedError:
